+++ src/core/MeadeParserDistance.cpp
@@ -1,0 +1,26 @@
+/**
+ * @file MeadeParserDistance.cpp
+ * @brief Distance-family (`:D...`) dispatcher for the Meade LX200 parser.
+ */
+
+#include "core/MeadeParser.hpp"
+#include "core/MeadeParserHelpers.hpp"
+
+namespace oat
+{
+namespace core
+{
+namespace meade
+{
+
+MeadeResponse handleMeadeDistance(const char *, IMeadeDistanceHandlers &h)
+{
+    MeadeResponse r;
+    writeChar(r, h.onIsSlewingRaOrDec() ? '|' : ' ');
+    writeTerminator(r);
+    return r;
+}
+
+}  // namespace meade
+}  // namespace core
+}  // namespace oat
