+++ src/core/MeadeParserInit.cpp
@@ -1,0 +1,23 @@
+/**
+ * @file MeadeParserInit.cpp
+ * @brief Init-family (`:I...`) dispatcher for the Meade LX200 parser.
+ */
+
+#include "core/MeadeParser.hpp"
+
+namespace oat
+{
+namespace core
+{
+namespace meade
+{
+
+MeadeResponse handleMeadeInit(const char *, IMeadeInitHandlers &h)
+{
+    h.onEnterSerialControl();
+    return MeadeResponse {};
+}
+
+}  // namespace meade
+}  // namespace core
+}  // namespace oat
